feat!: Improve hashing performance and security (#33)
## Problem 1: Performance
fictional's job is to map inputs to outputs, in a way that is deterministic, and ideally difficult to reverse. The right tool for the job here is hash functions - to map each input to a value that we can use as in index to an entry a data set, or from which we can generate an output deterministically.

Fictional currently makes use of md5 to do this. The issue with md5 for us is performance - because of how heavily we need to rely on hashing, the faster we can hash the better. md5 does fairly well compared to the hashing functions it is usually compared to (e.g. sha2), but can still make some use cases infeasibly slow.

Part of the performance hit is also due to the fact that we need to convert the md5 result from a large hex value, into a small enough integer to be usable as an index. At the same time, it is also overkill, since in this case the limiting factor impacting collision resistance is not the hashing function itself, but rather the data set size.

See [copycat#36](https://github.com/snaplet/copycat/pull/36) for benchmarks and more context.

## Problem 2: Security
fictional supports [salt](https://en.wikipedia.org/wiki/Salt_(cryptography)) as a way of making it more difficult to reverse engineer inputs. However, the use cases for fictional are often ones where per-item salt isn't easily possible (for example, deterministically transforming an already existing dataset - in contrast to, for example, a password storage use case where there's a natural opportunity to randomly generate and store per-item salt alongside each item).

As a result, we're basically using the salt as a secret key. Using a secret key when hashing is not problematic, its just that this isn't really the purpose of salt, so there is no design intent for making key recovery difficult. The right tool for the job here is a keyed hashing function - the algorithm was designed with this in mind.

## Solution
Use [SipHash](https://en.wikipedia.org/wiki/SipHash)
* It is designed to make both recovery of the key and the input value difficult
* It has performance comparable to non-cryptographic hash functions
* It appears to perform well for the properties important to us, including:
  * [Diffusion](https://en.wikipedia.org/wiki/Confusion_and_diffusion): Ideally, any change in the hash input should cause a drastic enough change in the hash output - so that an attacker is not able to infer anything about the input
  * Distribution: The hash output values should distribute as evenly as possible across the hash space

## Other links
* Corresponding copycat PR: https://github.com/snaplet/copycat/pull/36
* Video with context: https://www.loom.com/share/489d8300e232435987dbb68fdc1704b9